engine/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pyswisseph==2.10.3.2
python-dotenv==1.1.0
pytest==8.3.5
httpx==0.28.1
pytest-asyncio==0.25.3
tzdata==2025.2
timezonefinder==6.5.0
